loplugin:redundantcast
...after 11b15571475414ef853e21a6c96afa2ac81f848f "convert KernArray from
sal_Int32 to double"

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176052
Tested-by: Jenkins

+++ svx/source/svdraw/svdotextdecomposition.cxx
@@ -299,7 +299,7 @@ namespace
             aDXArray.reserve(rInfo.mnTextLen);
             for(sal_Int32 a=0; a < rInfo.mnTextLen; a++)
             {
-                aDXArray.push_back(static_cast<double>(rInfo.mpDXArray[a]));
+                aDXArray.push_back(rInfo.mpDXArray[a]);
             }
         }
